Organize component files into app and shared directories
