+++ client/src/components/app/InfoHeader.vue
