+++ client/src/components/app/ProgressBar.vue
