+++ client/src/components/shared/SideNav.vue
