+++ client/src/components/shared/Tip.vue
