+++ client/src/views/Login.vue
@@ -49,8 +49,8 @@
 
 <script>
 import { mapActions } from 'vuex';
-import Alert from '@/components/Alert';
-import SideNav from '@/components/SideNav';
+import Alert from '@/components/shared/Alert';
+import SideNav from '@/components/shared/SideNav';
 
 export default {
   components: {
